(ci) Actually compile the mem_stats platform arms, against stubs
The arm check could only say that an arm had not picked up another
arm's code; an arm whose own header was missing or wrong was recorded as
"SDK header absent" and passed.  Three console failures reached CI in
exactly that shape - 3ds.h never included at all, gccore.h guessed at,
and SYSMEM1_SIZE taken from the SDK when it is RetroArch's own - and the
check waved all three through.

The SDKs are not on a runner, but they do not have to be.  Each arm
names a handful of symbols, so give it a stub header declaring those and
compile it for real: thirteen headers, none longer than a dozen lines.
Warnings fail the arm as well as errors, since a diagnostic is the whole
signal here.

What this proves and what it does not: the arm's own code is well formed
and every symbol it uses is one it arranged to have.  It cannot prove
the real SDK matches the stub - a stub that is wrong in the same
direction as the code would agree with it.  That is a smaller gap than
the one it closes.

Writing the stubs immediately turned up a real defect in the PS3 arm,
which is the argument for them: it cast through u64, a psl1ght typedef
the frontend had by way of its own includes and this file never
declared.  The PS3 job could not have caught it, because that build was
still failing at the link stage.  It casts through uint64_t now, which
needs nothing declaring.

Checked by reintroducing each of the three failures that reached CI:
removing the gx_defines.h include fails gamecube, removing 3ds/os.h
fails 3ds, restoring the u64 cast fails ps3.  All fifteen arms pass with
them restored - the two Windows variants and both Apple ones included,
which were never compiled anywhere before now.

+++ libretro-common/memory/mem_stats.c
@@ -295,7 +295,7 @@ uint64_t mem_stats_total(void)
 #elif defined(MEM_STATS_PS3)
    {
       sys_memory_info_t mem_info;
-      sys_memory_get_user_memory_size((u64)&mem_info);
+      sys_memory_get_user_memory_size((uint64_t)(uintptr_t)&mem_info);
       return (uint64_t)mem_info.total;
    }
 #elif defined(MEM_STATS_EMSCRIPTEN)
@@ -419,7 +419,7 @@ uint64_t mem_stats_free(void)
    /* named get_mem_used in the frontend, but what it read was avail */
    {
       sys_memory_info_t mem_info;
-      sys_memory_get_user_memory_size((u64)&mem_info);
+      sys_memory_get_user_memory_size((uint64_t)(uintptr_t)&mem_info);
       return (uint64_t)mem_info.avail;
    }
 #elif defined(MEM_STATS_EMSCRIPTEN)
